+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Updates `getFeaturesByCoordinate` in `wmsFeatureProvider` to consider the extent and to respect the `tilingSchema` in the bbox calculation
 - adds `htmlPositionFeatureTitle` to `wmsFeatureProvider` which is added as to the point feature created if (responseType === 'text/html').
 - Fixes an issue with COG layers, where there would be stripes in 3D.
+- Added option `sourceOptions` to COGLayer to forward options to the Openlayers source: https://openlayers.org/en/latest/apidoc/module-ol_source_GeoTIFF-GeoTIFFSource.html
 
 ### 6.2.2